requirements.txt
absl-py==1.4.0
astunparse==1.6.3
cachetools==5.3.0
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
coloredlogs==15.0.1
contourpy==1.0.7
cycler==0.10.0
filelock==3.12.0
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
flatbuffers==23.3.3
fonttools==4.39.3
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.8.0
humanfriendly==10.0
idna==2.10
imageio==2.28.0
itsdangerous==2.1.2
jax==0.4.8
Jinja2==3.1.2
keras==2.12.0
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
ml-dtypes==0.1.0
mpmath==1.3.0
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
onnx==1.13.1
onnxruntime==1.14.1
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
Pillow==9.5.0
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.0
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
roboflow==1.0.5
rsa==4.9
scikit-image==0.20.0
scipy==1.10.1
six==1.16.0
SQLAlchemy==2.0.11
supervision==0.6.0
sympy==1.11.1
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.12.0
tensorflow-macos==2.12.0
termcolor==2.3.0
tifffile==2023.4.12
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
wget==3.2
wrapt==1.14.1
WTForms==3.0.1
